Avoid `use cache`
`use cache` currently demands a suspense boundary when the `revalidate`
time is 0 or the `expire` time is < 300.  However, using a suspense
boundary causes rendering to be streamed, meaning the initial HTML
payload will not match what the benchmark expects.  Furthermore,
`cacheLife` causes a 500 error in production when the `revalidate`
option is < 1.  Therefore, avoid `use cache` until the API is mature.